+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## New Features in This Release
+- feat(Application): add split array
 - feat(Application): add replacing separators by comma
 - feat(Application): add catching custom separator
 - feat: setup project